pyproject.toml
[tool.poetry]
name = "distance-education-system"
version = "0.1.0"
description = ""
authors = ["temerk1n <nikita.verigin23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.0"
django-compat = "^1.0.15"
apscheduler = "^3.10.4"
drf-spectacular = "^0.27.1"
django-rest-enumfield = "^0.2.0"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
